pyproject.toml
[tool.poetry]
name = "youtube-summarizer"
version = "0.1.0"
description = ""
authors = ["HenriqueCarniel <henriquecarniel80@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.156.0"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.12.23"
tdqm = "^0.0.1"
openai-whisper = "^20240930"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
